--- frontend/index.html
@@ -1,51 +1,0 @@
-﻿<!DOCTYPE html>
-<!-- https://console.cloud.google.com/ -->
-
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>Chameleon EVCFLO Project</title>
-    <meta name="description" content="Chameleon EVCFLO project data demonstration">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://storage.cloud.google.com/evcflo_static_frontend/style.css">
-</head>
-
-<body>
-    <h1>Chameleon Electric Vehicle Charger Forecasting and Location Optimization project</h1>
-    <h2>Google map API demo</h2>
-
-    <!--Nav strip -->
-    <hr>
-    
-    <!--LOCAL-->
-    <!-- <a href="./index.html"> | Home.</a>
-    <a href="./add_station.html"> | Add a Charge Station here.</a>
-    <a href="./add_suggested_station.html"> | Add a Suggested Station here.</a>
-    <a href="./add_bounding.html"> | Add a Bounding Box here.</a>
-    <a href="./notebooks.html"> | View Research.</a>
-    <a href="https://github.com/Chameleon-company/EVCFLO"> | GitHub. |</a>
-    <a href="http://localhost:5000/frontend/visualizations.html"> | Visualizations. |</a>
-    -->
-
-    <!--LIVE-->
-    
-    <a href="https://storage.googleapis.com/evcflo_static_frontend/index.html"> | Home.</a>
-    <a href="https://storage.cloud.google.com/evcflo_static_frontend/add_station.html"> | Add a Charge Station here.</a>
-    <a href="https://storage.cloud.google.com/evcflo_static_frontend/add_suggested_station.html"> | Add a Suggested Station here.</a>
-    <a href="https://storage.cloud.google.com/evcflo_static_frontend/add_bounding.html"> | Add a Bounding Box here.</a>    
-    <a href="https://storage.cloud.google.com/evcflo_static_frontend/notebooks.html"> | View Research.</a>
-    <a href="https://github.com/Chameleon-company/EVCFLO"> | GitHub. |</a>
-    <a href ="https://storage.cloud.google.com/evcflo_static_frontend/visualizations.html"> | Visualizations. |</a>
-    <hr>
-
-    <!--The div element for the map -->
-    <div id="map"></div>
-
-    <script src="https://storage.cloud.google.com/evcflo_static_frontend/main.js"></script>
-    <!-- <script src="./main.js"></script> -->
-    <script src="https://maps.googleapis.com/maps/api/js?key=&MAPS_KEY&callback=initMap&v=weekly"></script>
-</body>
-
-</html>
