+++ frontend/oldfrontend/add_station.html
@@ -1,0 +1,105 @@
+﻿<!DOCTYPE html>
+
+<html>
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <title></title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://storage.cloud.google.com/evcflo_static_frontend/style.css">
+</head>
+
+<body>
+    <h1>Chameleon Electric Vehicle Charger Forecasting and Location Optimization project</h1>
+    <h2>Add a charger station to the database</h2>
+
+    <!--Nav strip -->
+    <hr>
+    <!--LOCAL-->
+    <!--    
+    <a href="http://localhost:5000/frontend/index.html"> | Home.</a>
+    <a href="http://localhost:5000/frontend/add_station.html"> | Add a Charge Station here.</a>
+    <a href="http://localhost:5000/frontend/add_suggested_station.html"> | Add a Suggested Station here.</a>
+    <a href="http://localhost:5000/frontend/add_bounding.html"> | Add a Bounding Box here.</a>
+    <a href="http://localhost:5000/frontend/notebooks.html"> | View Research.</a>
+    <a href="http://localhost:5000/frontend/visualizations.html"> | Visualizations. |</a>
+
+    -->
+    <!--LIVE-->
+    <a href="https://storage.googleapis.com/evcflo_static_frontend/index.html"> | Home.</a>
+    <a href="https://storage.cloud.google.com/evcflo_static_frontend/add_station.html"> | Add a Charge Station here.</a>
+    <a href="https://storage.cloud.google.com/evcflo_static_frontend/add_suggested_station.html"> | Add a Suggested Station here.</a>
+    <a href="https://storage.cloud.google.com/evcflo_static_frontend/add_bounding.html"> | Add a Bounding Box here.</a>    
+    <a href="https://storage.cloud.google.com/evcflo_static_frontend/notebooks.html"> | View Research.</a>
+    <a href="https://github.com/Chameleon-company/EVCFLO"> | GitHub. |</a>
+    <a href ="https://storage.cloud.google.com/evcflo_static_frontend/visualizations.html"> | Visualizations. |</a>
+    <hr>
+
+    <form action="https://server-e657pcuziq-ts.a.run.app/api/add_station" method="POST">
+        <label for="name">Name:</label>
+        <input type="text" id="name" name="name"><br>
+        <label for="address">Address:</label>
+        <input type="text" id="address" name="address"><br>
+        <label for="longitude">Longitude:</label>
+        <input type="number" step="0.0001" id="longitude" name="longitude"><br>
+        <label for="latitude">Latitude:</label>
+        <input type="number" step="0.0001" id="latitude" name="latitude"><br>
+        <label for="location_category">Location Category:</label>
+        <input type="text" id="location_category" name="location_category"><br>
+        <label for="location_sub_category">Location Subcategory:</label>
+        <input type="text" id="location_sub_category" name="location_sub_category"><br>
+        <label for="description">Description:</label>
+        <input type="text" id="description" name="description"><br>
+        <label for="open_date">Open Date:</label>
+        <input type="text" id="open_date" name="open_date"><br>
+        <label for="parking">Pricing:</label>
+        <input type="text" id="parking" name="parking"><br>
+        <label for="pricing">Pricing:</label>
+        <input type="text" id="pricing" name="pricing"><br>
+        <label for="free_use">Free Use:</label>
+        <input type="text" id="free_use" name="free_use"><br>
+        <label for="contact">Contact Details:</label>
+        <input type="text" id="contact" name="contact"><br>
+        <label for="networks">Networks (Format: []):</label>
+        <input type="text" id="networks" name="networks"><br>
+        <label for="total_plugs">Total Plugs:</label>
+        <input type="number" id="total_plugs" name="total_plugs"><br>
+        <label for="port_level_type">Port Level Type:</label>
+        <input type="text" id="port_level_type" name="port_level_type"><br>
+        <label for="power_outputs_kw">Power Output (kW, Format: []):</label>
+        <input type="number" id="power_outputs_kw" name="power_outputs_kw"><br>
+        <label for="renewable_power_supply">Renewable Power Supply:</label>
+        <input type="text" id="renewable_power_supply" name="renewable_power_supply"><br>
+        <label for="Plugs_CHAdeMO">Number of plugs - CHAdeMO:</label>
+        <input type="number" id="Plugs_CHAdeMO" name="Plugs_CHAdeMO"><br>
+        <label for="Plugs_Tesla">Number of plugs - Tesla:</label>
+        <input type="number" id="Plugs_Tesla" name="Plugs_Tesla"><br>
+        <label for="Plugs_CCS_SAE">Number of plugs - CCS_SAE:</label>
+        <input type="number" id="Plugs_CCS_SAE" name="Plugs_CCS_SAE"><br>
+        <label for="Plugs_J1772">Number of plugs - J1772:</label>
+        <input type="number" id="Plugs_J1772" name="Plugs_J1772"><br>
+        <label for="Plugs_Type2">Number of plugs - Type2:</label>
+        <input type="number" id="Plugs_Type2" name="Plugs_Type2"><br>
+        <label for="Plugs_Three_Phase">Number of plugs - Three Phase:</label>
+        <input type="number" id="Plugs_Three_Phase" name="Plugs_Three_Phase"><br>
+        <label for="Plugs_Commando">Number of plugs - Commando:</label>
+        <input type="number" id="Plugs_Commando" name="Plugs_Commando"><br>
+        <label for="Plugs_Wall_AU_NZ">Number of plugs - Wall_AU:</label>
+        <input type="number" id="Plugs_Wall_AU_NZ" name="Plugs_Wall_AU_NZ"><br>
+        <label for="Plugs_Caravan_Mains">Number of plugs - Caravan_Mains:</label>
+        <input type="number" id="Plugs_Caravan_Mains" name="Plugs_Caravan_Mains"><br>
+        <label for="Plugs_Other">Number of plugs - Other:</label>
+        <input type="number" id="Plugs_Other" name="Plugs_Other"><br>
+        <label for="source">Source of Datapoint:</label>
+        <input type="text" id="source" name="source"><br>
+        <label for="source_date">Source Date:</label>
+        <input type="date" id="source_date" name="source_date"><br>
+        <input type="submit" value="Submit">
+    </form>
+
+    <script src="" async defer></script>
+</body>
+
+</html>
