+++ frontend/oldfrontend/notebooks.html
@@ -1,0 +1,89 @@
+﻿<!DOCTYPE html>
+
+<html>
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <title></title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://storage.cloud.google.com/evcflo_static_frontend/style.css">
+</head>
+
+<body>
+    <h1>Chameleon Electric Vehicle Charger Forecasting and Location Optimization project</h1>
+    <h2>Data Analysis Reports</h2>
+
+    <!--Nav strip -->
+    <hr>
+    <!--LOCAL-->
+    <!--    
+    <a href="http://localhost:5000/frontend/index.html"> | Home.</a>
+    <a href="http://localhost:5000/frontend/add_station.html"> | Add a Charge Station here.</a>
+    <a href="http://localhost:5000/frontend/add_suggested_station.html"> | Add a Suggested Station here.</a>
+    <a href="http://localhost:5000/frontend/add_bounding.html"> | Add a Bounding Box here.</a>
+    <a href="http://localhost:5000/frontend/notebooks.html"> | View Research.</a>
+    <a href="http://localhost:5000/frontend/visualizations.html"> | Visualizations. |</a>
+    -->
+    <!--LIVE-->
+    <a href="https://storage.googleapis.com/evcflo_static_frontend/index.html"> | Home.</a>
+    <a href="https://storage.cloud.google.com/evcflo_static_frontend/add_station.html"> | Add a Charge Station here.</a>
+    <a href="https://storage.cloud.google.com/evcflo_static_frontend/add_suggested_station.html"> | Add a Suggested Station here.</a>
+    <a href="https://storage.cloud.google.com/evcflo_static_frontend/add_bounding.html"> | Add a Bounding Box here.</a>
+    <a href="https://storage.cloud.google.com/evcflo_static_frontend/notebooks.html"> | View Research.</a>
+    <a href="https://github.com/Chameleon-company/EVCFLO"> | GitHub. |</a>
+    <a href ="https://storage.cloud.google.com/evcflo_static_frontend/visualizations.html"> | Visualizations. |</a>
+    <hr>
+
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EVCS_Turku-Finland_Analysis.html">EVCS Turku-Finland Analysis</a></h3>
+    <p>Examination of charge duration and energy consumption on sample Turku Finland dataset.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EVData(EDA_Analysis).html">EVData(EDA_Analysis)</a></h3>
+    <p>Investigates the characteristics of existing charging stations in Queensland and Melbourne Austrlia.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EVData(Google_API).html">EVData(Google_API)</a></h3>
+    <p>Utilizing Google Maps API in order to investigate landmark count in close proximity to charging stations found in Queensland and Melbourne Australia.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EV_Registered_Data.html">EV_Registered_Data</a></h3>
+    <p>Studies the volume of Electric Vehicles that have been registered globally, with focused interest in Australia and the US.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EVCS_Boulder_Colorado%20Analysis.html">EVCS_Boulder_Colorado Analysis</a></h3>
+    <p>Analysis and visualisation of EV charging station energy consumption data for charging station in Boulder, Colorado.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EVCS%20-%20Consumption%20Prediction%20by%20Time%20Series.html">EVCS - Consumption Prediction by Time Series</a></h3>
+    <p>In this analysis, we would like to develop a prediction model based on time series data generated from the EVCS dataset in the UK and forecast future energy consumption trends for the next 365 days. Our analysis contains four sections: Data Cleaning,
+        Data Exploration, Pattern Recognition, and Prediction Model. The results of evaluation metrics suggest that the model has strong predictive power, while future work directions are also pointed out.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EVCS%20-%20Consumption%20Prediction%20by%20Transaction%20Records.html">EVCS - Consumption Prediction by Transaction Records</a></h3>
+    <p>In this analysis, we will build a Machine Learning model using a Random Forest algorithm based on an EVCS dataset in Netherland to explore the relationships between energy consumption and all other features. We also provide graphical presentations
+        of information extracted from our research. The analysis has four main sections: Data Cleaning, Data Exploration, Feature Selection, and Prediction Model. Our final model returns an accuracy of 87.33% which provides great prediction power by individual
+        transaction records.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EVCS%20Visualisation%20-%20Barcelona.html">EVCS Visualisation - Barcelona</a></h3>
+    <p>In this analysis, we would like to explore the dataset of EVCS in Barcelona Spain by investigating various important features and their relationships. We use various distribution methods to fit in the key attributes to explore the patterns, while
+        providing other visualised methods to illustrate our findings.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/AustinEVdata.html">AustinEVdata</a></h3>
+    <p>Investigation and cleaning a dataset of EV charging stations in Austin, Texas.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/AustinEVData(Google_Maps_API).html">AustinEVData(Google_Maps_API)</a></h3>
+    <p>Utilizing Google Maps API in order to investigate landmark count in close proximity to charging stations found in Austin, Texas.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/AustinEV_Clustering.html">AustinEV_Clustering</a></h3>
+    <p>Clustering analysis of the landmark proximity dataset for Austin, Texas.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EVCS - Data Cleaning and Analysis - Cary (NC).html">Analysis of Cary, N.Carolina USA</a></h3>
+    <p>Data cleaning and analysis of EV vehicle data from Cary, N.Carolina, USA. Includes analysis of charging time and energy usage.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/ukEVlocations.html">UK EV Charging Locations</a></h3>
+    <p>Analysis of UK EV Charging Stations.</p>
+    <hr>
+    <h3><a href="https://storage.googleapis.com/evcflo_static_frontend/notebooks/EVCS Interactive Map-melbourne and geelong.html">EVCS - Interactive Map Development - Melbourne and Geelong Region</a></h3>
+    <p>In this analysis, we will built an Interactive Map that provides indications of exisiting EVCS locations, as well as the possible future charging sites. Techniques used in this analysis is based on the EVCS location datasets of Melbourne and Geelong, and Melbourne off street parking lot location dataset.</p>
+    <hr>
+
+    <script src="" async defer></script>
+</body>
+
+</html>
