+++ frontend/oldfrontend/style.css
@@ -1,0 +1,17 @@
+#map {
+    /* Setup the size of the map */
+    height: 600px;
+    width: 100%;
+}
+
+
+/* Form styling for adding charger station... */
+
+label {
+    cursor: pointer;
+    display: inline-block;
+    padding: 3px 6px;
+    text-align: right;
+    width: 230px;
+    vertical-align: top;
+}
